backend/requirements.txt
fastapi==0.111.1
uvicorn==0.25.0
sqlalchemy==2.0.25
pydantic==2.6.2
python-multipart==0.0.7
passlib[bcrypt]==1.7.4
python-jose==3.3.0
itsdangerous==2.1.2
email-validator>=2.3.0
python-dotenv==1.0.0

# Data processing / documents
pandas>=2.1.1
openpyxl>=3.1.0
PyPDF2>=3.0.0
python-docx>=0.8.12
docx2txt>=0.8
python-pptx>=0.6.21
pillow>=10.0.0
pytesseract>=0.3.10

# OpenAI API
openai>=1.0.0

# Optional helpers (if needed for Excel/CSV)
xlrd>=2.0.1
